src/main/java/ru/serg_nik/foodvoice/service/VoiceService.java
package ru.serg_nik.foodvoice.service;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ru.serg_nik.foodvoice.dto.VoiceDto;
import ru.serg_nik.foodvoice.model.Voice;
import ru.serg_nik.foodvoice.repository.MenuRepository;
import ru.serg_nik.foodvoice.repository.UserRepository;
import ru.serg_nik.foodvoice.repository.VoiceRepository;

import java.time.LocalDate;
import java.time.LocalTime;
import java.util.UUID;
import java.util.concurrent.TimeoutException;

@Service
@Slf4j
public class VoiceService extends BaseEntityService<Voice, VoiceRepository> {

    public static final LocalTime VOICE_CHANGE_STOP_TIME = LocalTime.of(11, 0);

    private final UserRepository userRepository;
    private final MenuRepository menuRepository;

    @Autowired
    public VoiceService(VoiceRepository repository, UserRepository userRepository, MenuRepository menuRepository) {
        super(repository);
        this.userRepository = userRepository;
        this.menuRepository = menuRepository;
    }

    @Override
    protected void prepareBeforeCreate(Voice entity) {
        entity.setDate(LocalDate.now());
    }

    @Override
    protected void prepareBeforeUpdate(Voice entity) {
    }

    public VoiceDto vote(UUID userId, UUID menuId) {
        Voice voice = new Voice();
        voice.setUser(userRepository.getOne(userId));
        voice.setMenu(menuRepository.getOne(menuId));
        return new VoiceDto(super.create(voice));
    }

    public VoiceDto changeVote(UUID id, UUID menuId) throws TimeoutException {
        if (LocalTime.now().isAfter(VOICE_CHANGE_STOP_TIME)) {
            throw new TimeoutException("Истекло время, до которого можно было изменить выбор ресторана");
        }
        Voice voice = get(id);
        if (voice.getId().equals(id)) {
            return new VoiceDto(voice);
        } else {
            voice.setMenu(menuRepository.getOne(menuId));
            return new VoiceDto(repository.save(voice));
        }
    }

}
